+++ src/DevilDaggersInfo.Tools/Ui/ReplayEditor/Events/EventType.cs
@@ -18,6 +18,8 @@ public enum EventType
 	Hit,
 	Transmute,
 
+	InitialInputs,
+	Inputs,
 	Death,
 	End,
 }